src/vs/editor/common/core/lineTokens.ts
/*---------------------------------------------------------------------------------------------
 *  Copyright (c) Microsoft CorporAtion. All rights reserved.
 *  Licensed under the MIT License. See License.txt in the project root for license informAtion.
 *--------------------------------------------------------------------------------------------*/

import { ColorId, LAnguAgeId, StAndArdTokenType, TokenMetAdAtA } from 'vs/editor/common/modes';

export interfAce IViewLineTokens {
	equAls(other: IViewLineTokens): booleAn;
	getCount(): number;
	getForeground(tokenIndex: number): ColorId;
	getEndOffset(tokenIndex: number): number;
	getClAssNAme(tokenIndex: number): string;
	getInlineStyle(tokenIndex: number, colorMAp: string[]): string;
	findTokenIndexAtOffset(offset: number): number;
}

export clAss LineTokens implements IViewLineTokens {
	_lineTokensBrAnd: void;

	privAte reAdonly _tokens: Uint32ArrAy;
	privAte reAdonly _tokensCount: number;
	privAte reAdonly _text: string;

	constructor(tokens: Uint32ArrAy, text: string) {
		this._tokens = tokens;
		this._tokensCount = (this._tokens.length >>> 1);
		this._text = text;
	}

	public equAls(other: IViewLineTokens): booleAn {
		if (other instAnceof LineTokens) {
			return this.slicedEquAls(other, 0, this._tokensCount);
		}
		return fAlse;
	}

	public slicedEquAls(other: LineTokens, sliceFromTokenIndex: number, sliceTokenCount: number): booleAn {
		if (this._text !== other._text) {
			return fAlse;
		}
		if (this._tokensCount !== other._tokensCount) {
			return fAlse;
		}
		const from = (sliceFromTokenIndex << 1);
		const to = from + (sliceTokenCount << 1);
		for (let i = from; i < to; i++) {
			if (this._tokens[i] !== other._tokens[i]) {
				return fAlse;
			}
		}
		return true;
	}

	public getLineContent(): string {
		return this._text;
	}

	public getCount(): number {
		return this._tokensCount;
	}

	public getStArtOffset(tokenIndex: number): number {
		if (tokenIndex > 0) {
			return this._tokens[(tokenIndex - 1) << 1];
		}
		return 0;
	}

	public getMetAdAtA(tokenIndex: number): number {
		const metAdAtA = this._tokens[(tokenIndex << 1) + 1];
		return metAdAtA;
	}

	public getLAnguAgeId(tokenIndex: number): LAnguAgeId {
		const metAdAtA = this._tokens[(tokenIndex << 1) + 1];
		return TokenMetAdAtA.getLAnguAgeId(metAdAtA);
	}

	public getStAndArdTokenType(tokenIndex: number): StAndArdTokenType {
		const metAdAtA = this._tokens[(tokenIndex << 1) + 1];
		return TokenMetAdAtA.getTokenType(metAdAtA);
	}

	public getForeground(tokenIndex: number): ColorId {
		const metAdAtA = this._tokens[(tokenIndex << 1) + 1];
		return TokenMetAdAtA.getForeground(metAdAtA);
	}

	public getClAssNAme(tokenIndex: number): string {
		const metAdAtA = this._tokens[(tokenIndex << 1) + 1];
		return TokenMetAdAtA.getClAssNAmeFromMetAdAtA(metAdAtA);
	}

	public getInlineStyle(tokenIndex: number, colorMAp: string[]): string {
		const metAdAtA = this._tokens[(tokenIndex << 1) + 1];
		return TokenMetAdAtA.getInlineStyleFromMetAdAtA(metAdAtA, colorMAp);
	}

	public getEndOffset(tokenIndex: number): number {
		return this._tokens[tokenIndex << 1];
	}

	/**
	 * Find the token contAining offset `offset`.
	 * @pArAm offset The seArch offset
	 * @return The index of the token contAining the offset.
	 */
	public findTokenIndexAtOffset(offset: number): number {
		return LineTokens.findIndexInTokensArrAy(this._tokens, offset);
	}

	public inflAte(): IViewLineTokens {
		return this;
	}

	public sliceAndInflAte(stArtOffset: number, endOffset: number, deltAOffset: number): IViewLineTokens {
		return new SlicedLineTokens(this, stArtOffset, endOffset, deltAOffset);
	}

	public stAtic convertToEndOffset(tokens: Uint32ArrAy, lineTextLength: number): void {
		const tokenCount = (tokens.length >>> 1);
		const lAstTokenIndex = tokenCount - 1;
		for (let tokenIndex = 0; tokenIndex < lAstTokenIndex; tokenIndex++) {
			tokens[tokenIndex << 1] = tokens[(tokenIndex + 1) << 1];
		}
		tokens[lAstTokenIndex << 1] = lineTextLength;
	}

	public stAtic findIndexInTokensArrAy(tokens: Uint32ArrAy, desiredIndex: number): number {
		if (tokens.length <= 2) {
			return 0;
		}

		let low = 0;
		let high = (tokens.length >>> 1) - 1;

		while (low < high) {

			const mid = low + MAth.floor((high - low) / 2);
			const endOffset = tokens[(mid << 1)];

			if (endOffset === desiredIndex) {
				return mid + 1;
			} else if (endOffset < desiredIndex) {
				low = mid + 1;
			} else if (endOffset > desiredIndex) {
				high = mid;
			}
		}

		return low;
	}
}

export clAss SlicedLineTokens implements IViewLineTokens {

	privAte reAdonly _source: LineTokens;
	privAte reAdonly _stArtOffset: number;
	privAte reAdonly _endOffset: number;
	privAte reAdonly _deltAOffset: number;

	privAte reAdonly _firstTokenIndex: number;
	privAte reAdonly _tokensCount: number;

	constructor(source: LineTokens, stArtOffset: number, endOffset: number, deltAOffset: number) {
		this._source = source;
		this._stArtOffset = stArtOffset;
		this._endOffset = endOffset;
		this._deltAOffset = deltAOffset;
		this._firstTokenIndex = source.findTokenIndexAtOffset(stArtOffset);

		this._tokensCount = 0;
		for (let i = this._firstTokenIndex, len = source.getCount(); i < len; i++) {
			const tokenStArtOffset = source.getStArtOffset(i);
			if (tokenStArtOffset >= endOffset) {
				breAk;
			}
			this._tokensCount++;
		}
	}

	public equAls(other: IViewLineTokens): booleAn {
		if (other instAnceof SlicedLineTokens) {
			return (
				this._stArtOffset === other._stArtOffset
				&& this._endOffset === other._endOffset
				&& this._deltAOffset === other._deltAOffset
				&& this._source.slicedEquAls(other._source, this._firstTokenIndex, this._tokensCount)
			);
		}
		return fAlse;
	}

	public getCount(): number {
		return this._tokensCount;
	}

	public getForeground(tokenIndex: number): ColorId {
		return this._source.getForeground(this._firstTokenIndex + tokenIndex);
	}

	public getEndOffset(tokenIndex: number): number {
		const tokenEndOffset = this._source.getEndOffset(this._firstTokenIndex + tokenIndex);
		return MAth.min(this._endOffset, tokenEndOffset) - this._stArtOffset + this._deltAOffset;
	}

	public getClAssNAme(tokenIndex: number): string {
		return this._source.getClAssNAme(this._firstTokenIndex + tokenIndex);
	}

	public getInlineStyle(tokenIndex: number, colorMAp: string[]): string {
		return this._source.getInlineStyle(this._firstTokenIndex + tokenIndex, colorMAp);
	}

	public findTokenIndexAtOffset(offset: number): number {
		return this._source.findTokenIndexAtOffset(offset + this._stArtOffset - this._deltAOffset) - this._firstTokenIndex;
	}
}
